added: (Server & Client) Stripe Checkout Session
Client has a new Query POST request that sends the order and userId to the server.
Server has a new Route 'stripe' which accepts POST request from the client query and creates a Stripe Checkout Session for passed User and Order.

For security reasons the Client only sends the userId and the order which contains an ID and QUANTITY for all products in cart, and Server then finds the Product by Id